[DOCS] Fix problem with Edit in github link
The link is created with docs in it, even if the mkdocs.yml file
mentions doc as the documentation folder. I am not aware if there is
any other dependency in the name doc, I have renamed it to docs and
changed it also in the mkdocs.yml file. It seems docs is hardcoded in
the readthedocs theme :(